moved polyfills into a separate file
